Add unwrap of json.Number
When unmarshalling JSON data, the standard behavior is to transform
numbers into floats. By telling the decoder to use Number, it will
instead produce json.Number instances. A json.Number can be queried
if it is a valid integer or a float. This commit ensures that the
wrap function is aware of json.Number and that it resolves it correctly
into an IntegerValue or FloatValue.