Remove dot imports
This commit removes dot imports across the project, the Go team
discourage their use outside of tests [1], with lint also flagging it.

It also gofmts the code.

Note the intentional blank import of functions as the last import in
pcore/entrypoint.go to mitigate segfault caused due to init ordering.

[1] https://github.com/golang/go/wiki/CodeReviewComments#import-dot