Change loader Entry to hold a Location instead of a string
Prior to this commit, the loader Entry would hold a string instead of
a Location. This forced locations to be resolved early (which means
generating the line index) rather than late (when printing an error).
This saves resources since the latter case only happens when something
is logged (warning or error).