Turn the Puppet Pcore instance into a proper singleton
Prior to this commit, it was possible to create several instances of
a Pcore and the last one created became the global `eval.Puppet`
instance. This commit changes this by replacing the method
`pcore.NewPcore()` with `pcore.InitializePuppet()`. The latter has no
return value and it is expected that users will instead use the global
`eval.Puppet` instance.