Move all state out of Evaluator and into Context
Before this commit, the evaluator held the definitions slice and a
default loader. This commit removes the need for the loader and moves
the definitions into the Context.

The primary reason for this refactoring is to clarify that the evaluator
can be a shared singleton whereas Context must be thread local.