Prevent iterator chaining causing lazy evaluation with incorrect stack
Evaluating iterator blocks lazily into arrays caused the block to be
evaluated after popping the ephemeral local scope. The lazy approach
also caused an unnecessary slowdown in the evaluator.

This commit ensures that iterators are resolved eagerly rather than
chained.

It will be necessary to revisit the topic of lazy evaluation when adding
support for lambdas. Such lambdas must be evaluated using the scope
that was in effect at the location where the lambda was declared.