Improve Pcore Do() and Process() to return error and take parent
This commit changes the Pcore Do() and Process() methods so that they
do not expose the panics that happens within the evaluator. Instead, the
well known panics are recovered and returned as errors.

Two new methods are also added to provide an alternative signature for
Do() and Process() where a parent Go context can be provided. The new
methods are DoWithParent() and ProcessWithParent().