Introduce Result, ResultSet and the previous node result $pnr variable
This commit changes how nodes created using relational operators are
evaluated so that each node now produces a Result or a ResultSet. The
results from previous node evaluations are also made availabel through
the global variable $pnr (should read as "previous nodes result").

The ApplyFunction API was changed so that it no longer uses handles.
Instead, it receives a slice of resources and is expected to return
a slice with the exact same size where each slot is ether an applied
resource or an ErrorObject.