Improve String formatting of Object and TypeSet
This commit moves enhances the logic used when producing the string
representation of Object and TypeSet types. An Object that is inside
of a TypeSet is now presented without the suffix that the TypeSet adds
to its name. It is also using the type alias form where the parent
is placed in front of the definition instead of inside.