Fix TypeSet bug causing failures for name using more than 2 segments
The loader logic for resolving typed names in a typeset was flawed and
didn't work correctly for names using more than 2 name segments. This
commit fixes that by introducing the methods `IsParent` and `RelativeTo`
to the `eval.TypedName` interface and then use the `RelativeTo` when
resolving types inside the typeset.