Ensure that an evaluated value can be used as a hash key.
Before this commit, an attempt to use an `_eval` construct as a key in
a hash would fail since the _eval could result in a concatentated string
expression. Such expressions must be allowed as hash keys although their
final value is unknown until the string gets evaluated.

Since the value isn't known until evaluation, it's not possible to fully
determine the uniqueness of the key. A best effort is used here where
the expression is transformed into its PN format. The string form of
that format is then used as key.