(GEN-7) Give special meaning to YAML key that is in the form of an array
This commit makes the Yaml2Ast converter sensitive to keys that are
in the form of an array so that the elements of the array forms a
functor (a receiver of a call parameter list) and ultimately results in
some kind of call.

If the array contains a name that starts with an uppercase character,
then that name is considered to be a type reference. A name starting
with a lowercase character is considered to be the name of a function
(if the name is the first in the array) or of a method.

The value the given key is converted to arguments for the call and can
be in the form of an array, a hash, or just a single value.

A single element array with a type name is special. It means that the
`new` method of that type should be called.