(GEN-12) Ensure that TypeSet loader resolves locally before parent
The TypeSet loader must resolve against its own types without namespace
prior to consulting the parent loader because the namespace is implicit
within the typeset. In other words, the type 'Foo' declared in a typeset
named 'Xyz' ('Xyz::Foo') must not be overridden by a parent 'Foo' when
refered to as 'Foo' within the typeset itself.