(GEN-15) Add ability for Node to evaluate lambda.
In order to pass functions and lambdas to the `sequential` and
`parallel` functions, the node must be able to use them as the thing
to evaluate. This commit adds a 'lambda' attribute to the node that,
if set, will cause the evaluation to be a single call to that lambda.