(LYR-162) Interpret Go struct pointer fields as Optional
This commit improves the eval.Reflect API so that an ObjectType created
from a go struct can have optional fields in the form of pointers. This
was not possible before because non-pointer values always have a "zero"
value in Go.

Since non-pointer values cannot hold a value of type `Undef`, this
commit also adds a check that such values are not declared as optional.

A new TaggedType is introduced to allow third party structures to be
amended with 'puppet' tags without needing to create copies of them.