(#14) Use interface rather than reflected pointer for serialization map
The Serializer/Deserializer would use an uintptr key derived from
`reflect.ValueOf(value).Pointer()` to keep track of values that had
previously been serialized to the stream. This turned out to be a bad
idea since some of the serialized values were generated and then
discarded which lead to the data being garbage collected and reused by
other objects, which of course then got the same pointer value.

This commit removes the use of Pointer values and instead uses the
serialized eval.Value interface directly as a key.

closes #14