Make StringValue and StringType more efficient and resource conservative
This commit replaces the `types.StringValue` struct with a type that
_is_  a string, i.e.:
```
type stringValue string
```
and adds `eval.StringValue` as an interface that this type implements.

The commit also breaks down the former types.StringType into three
different types. One for the unbounded String type, one for the size
constrained String type, and one for the value constrained String type
and adds the `eval.StringType` interface that these three types
implements.