Separate type resolution from AST evaluation
This commit adds a types.Resolve and types.ResolveWithParams method
that can resolve a type name or a type name with a set of parameters
into a corresponding type and removes the corresponding functionality
from the evaluator.

The commit also consistently change all names of meta types to
XxxMetaType instead of the former, non go-lint compliant Xxx_Type.

Relates to lyraproj/lyra#137