Make it possible to permanently change the loader of the eval.Context
In some situations, it is necessary to change the loader of the Context
permanently. The RPC plug-in where manifests are loaded using a file
based loader and then partly evaluated during a subsequent RPC call is
one such situation.

This commit adds the method SetLoader to the context.

Relates to lyraproj/lyra#148